Rename mesh-init -> control-plane
